requirements.txt
argcomplete==3.1.4
attrs==23.1.0
branca==0.6.0
certifi==2023.7.22
charset-normalizer==3.3.1
click==8.1.7
click-plugins==1.1.1
cligj==0.7.2
fiona==1.9.5
folium==0.14.0
geopandas==0.14.0
gtfs-functions==2.0.8
gtfs-realtime-bindings==1.0.0
idna==3.4
jenkspy==0.3.3
Jinja2==3.1.2
MarkupSafe==2.1.3
numpy==1.26.1
packaging==23.2
pandas==2.1.1
pendulum==2.1.2
plotly==5.17.0
protobuf==4.24.4
pyproj==3.6.1
python-dateutil==2.8.2
pytz==2023.3.post1
pytzdata==2020.1
PyYAML==6.0.1
requests==2.31.0
shapely==2.0.2
six==1.16.0
SQLAlchemy==2.0.22
tenacity==8.2.3
tomlkit==0.12.2
typing_extensions==4.8.0
tzdata==2023.3
unicode==2.9
urllib3==2.0.7
utm==0.7.0
xmltodict==0.13.0
yq==3.2.3
